requirements.txt
Django<3.0
djangorestframework
djangorestframework-datatables
djangorestframework-jwt
drf_yasg
psycopg2
